requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
torch
torchvision
python-Levenshtein
tqdm
argparse
typing